Added the base files for to do list application
These are the base files for a working implementation of the basic to do list app. Has the basics of creating a new group, and adding individual tasks to each of them. Also has basic task handling, including creating new tasks, checking them off, moving them within the list, and deleting the task.